+++ community/cypher/cypher-compiler-2.3/src/main/scala/org/neo4j/cypher/internal/compiler/v2_3/spi/QueryContext.scala
@@ -59,9 +59,9 @@ trait QueryContext extends TokenContext {
 
   def removeLabelsFromNode(node: Long, labelIds: Iterator[Int]): Int
 
-  def getPropertiesForNode(node: Long): Iterator[Long]
+  def getPropertiesForNode(node: Long): Iterator[Int]
 
-  def getPropertiesForRelationship(relId: Long): Iterator[Long]
+  def getPropertiesForRelationship(relId: Long): Iterator[Int]
 
   def getOrCreatePropertyKeyId(propertyKey: String): Int
